Revert "Restructured UI system files"
This reverts commit 5dcd1a7c36be310bb6e7043ea5e439e5f6c4e3d7.
