--- UISystem/FramerateUI.cpp
@@ -1,15 +1,0 @@
-#include "UISystem.h"
-#include <filesystem>        
-
-using namespace NCL;
-using namespace CSC8508;
-
-
-void UISystem::DisplayFramerate() {
-	ImGui::SetNextWindowPos(ImVec2(50, 50));
-	ImGui::SetNextWindowSize(ImVec2(120, 50));
-	bool open = true;
-	ImGui::Begin("Framerate", &open, ImGuiWindowFlags_NoResize);
-	ImGui::Text(std::to_string(1.0f / dt).c_str());
-	ImGui::End();
-}
