+++ UISystem/UISystem.cpp
@@ -69,3 +69,67 @@ void UISystem::DrawDemo() {
 		ImGui::ShowDemoWindow();
 	}
 }
+
+void UISystem::DisplayFramerate() {
+	ImGui::SetNextWindowPos(ImVec2(50, 50));
+	ImGui::SetNextWindowSize(ImVec2(120, 50));
+	bool open = true;
+	ImGui::Begin("Framerate", &open, ImGuiWindowFlags_NoResize);
+	ImGui::Text(std::to_string(1.0f / dt).c_str());
+	ImGui::End();
+}
+
+void UISystem::AudioSliders() {
+	ImGui::SetNextWindowPos(ImVec2(800, 300));
+	ImGui::SetNextWindowSize(ImVec2(500, 125));
+	ImGui::Begin("Audio Sliders");
+	
+	masterVolume = audioEngine->GetChannelVolume(ChannelGroupType::MASTER);
+	masterVolume = masterVolume * 100;
+	ImGui::SliderFloat("Master Volume", &masterVolume, 0, 100, "%.0f");
+	masterVolume = masterVolume / 100;
+	audioEngine->SetChannelVolume(ChannelGroupType::MASTER, masterVolume);
+
+	musicVolume = audioEngine->GetChannelVolume(ChannelGroupType::MUSIC);
+	musicVolume = musicVolume * 100;
+	ImGui::SliderFloat("Music Volume", &musicVolume, 0, 100, "%.0f");
+	musicVolume = musicVolume / 100;
+	audioEngine->SetChannelVolume(ChannelGroupType::MUSIC, musicVolume);
+
+	sfxVolume = audioEngine->GetChannelVolume(ChannelGroupType::SFX);
+	sfxVolume = sfxVolume * 100;
+	ImGui::SliderFloat("SFX Volume", &sfxVolume, 0, 100, "%.0f");
+	sfxVolume = sfxVolume / 100;
+	audioEngine->SetChannelVolume(ChannelGroupType::SFX, sfxVolume);
+
+	voiceVolume = audioEngine->GetChannelVolume(ChannelGroupType::CHAT);
+	voiceVolume = voiceVolume * 100;
+	ImGui::SliderFloat("Voice Volume", &voiceVolume, 0, 100, "%.0f");
+	voiceVolume = voiceVolume / 100;
+	audioEngine->SetChannelVolume(ChannelGroupType::CHAT, voiceVolume);
+
+	ImGui::End();
+}
+
+void UISystem::MainMenu() {
+	ImGui::SetNextWindowPos(ImVec2(80, 480));
+	ImGui::SetNextWindowSize(ImVec2(600, 500));
+	bool open = true;
+	ImGui::Begin("Main Menu", &open, ImGuiWindowFlags_NoTitleBar | ImGuiWindowFlags_NoBackground | ImGuiWindowFlags_NoResize);
+	ImGui::SetWindowFontScale(2);
+	if (ImGui::Button("Start Offline", ImVec2(600, 50))) {
+		uiList.remove(audioSliders);
+		menuOption = startOffline;
+	}
+	ImGui::SetCursorPosY(ImGui::GetCursorPosY() + 50);
+	if (ImGui::Button("Start as Server", ImVec2(600, 50))) {
+		uiList.remove(audioSliders);
+		menuOption = startServer;
+	}
+	ImGui::SetCursorPosY(ImGui::GetCursorPosY() + 50);
+	if (ImGui::Button("Start as Client", ImVec2(600, 50))) {
+		uiList.remove(audioSliders);
+		menuOption = StartClient;
+	}
+	ImGui::End();
+}
